requirements.txt
aiohttp==3.9.5
loguru==0.7.2
python-dotenv==1.0.1
tweepy==4.14.0
telethon==1.34.0
solana==0.33.0
web3==6.15.1
asyncpg==0.29.0
redis==5.0.3
openai==1.14.3
python-binance==1.0.19
base58
requests
websockets
solders
streamlit
plotly
pandas